Refactor antlr extensions
